pyproject.toml
[tool.poetry]
name = "migration-strawberry-test"
version = "0.1.0"
description = ""
authors = ["Alteian <Alteian@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = "^0.193.0"
strawberry-graphql-django = "^0.10.2"
django = "^4.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
